resnet50/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch
torchvision
pillow
txt 
flask
Flask-Injector
injector
waitress
gunicorn
prometheus_client
prometheus_flask_exporter
